reorganize record transforms and types
